Get a package's user-cache pins board

Returns a pkg's user-cache pins board. The board is automatically created if it doesn't already exist.

board(pkg, pkg_version = utils::packageVersion(pkg = pkg))

Arguments

pkg
Package name. A character scalar.

pkg_version
Optional version number of pkg to include in the board name. A package version number or NULL. If NULL, the board is agnostic about pkg's version number. Otherwise, a separate board is created for each pkg version number.

Value

The user-cache pins board belonging to pkg and pkg_version. An object ob class pins_board_folder.

See also