Delete a cached object from a package's user-cache pins board

Removes a cached object from a package's user-cache pins board.

rm_obj(board, id)

Arguments

board
Package's user-cache pins board.

id
Pin name uniquely identifying the object to be deleted from the pkg's user-cache pins board. A character scalar.

Value

id, invisibly.

Details

Note that it normally won't be necessary to delete a cached object using this function because cache_obj() will always overwrite a possibly existing object in the cache with the same id. If you want to ensure that no "data corpses" are left behind from using cache_obj(), consider clearing or even purging the whole package user-cache pins board at once.

See also