Cache the result of an expression from within a function

One-stop solution to add caching to an R function.

with_cache(
  expr,
  pkg,
  from_fn,
  ...,
  pkg_version = utils::packageVersion(pkg = pkg),
  use_cache = TRUE,
  cache_lifespan = "1 day"
)

Arguments

expr
Expression to cache.

pkg
Package name. A character scalar.

from_fn
Name of the function that expr is cached from, i.e. the name of the function that with_cache() is called from. A character scalar.

...
Arguments received by from_fn on which the caching should depend. This is fundamental to determine whether expr was already cached or not. The arguments must be specified unnamed (see examples). Dynamic dots are supported.

pkg_version
Optional version number of pkg to include when caching the result. A package version number or NULL. If not NULL, a separate user-cache pins board is created each time pkg's version number changes (e.g. after an upgrade), ensuring to never return cached results from a different (old) package version, irrespective of cache_lifespan.

use_cache
Return cached results if possible. If FALSE, results are always newly fetched regardless of cache_lifespan.

cache_lifespan
The duration after which cached results are refreshed (i.e. newly fetched). A valid lubridate duration. Only relevant if use_cache = TRUE.

Value

The result of evaluating expr, from cache if use_cache = TRUE and a cached result exists that hasn't exceeded cache_lifespan.

Details

expr is evaluated in the environment that with_cache() is called from.

See hash_fn_call() for details about how the pin name is constructed that uniquely identifies expr in the pkg's user-cache pins board.

Examples

# if the fn below would be part of a real package, we could instead define `this_pkg` globally
# using `this_pkg <- utils::packageName()`; instead, we now cache to pkgpins's cache (which
# itself never uses the cache)
this_pkg <- "pkgpins"

# let's define a fn that returns R pkg sys deps from cache
pkg_sys_deps <- function(pkg,
                         use_cache = TRUE,
                         cache_lifespan = "6h") {
  pkgpins::with_cache(
    expr = purrr::flatten(jsonlite::fromJSON(txt = paste0("https://sysreqs.r-hub.io/pkg/", pkg),
                                             simplifyVector = FALSE)),
    pkg = this_pkg,
    from_fn = "pkg_sys_deps",
    pkg,
    use_cache = use_cache,
    cache_lifespan = cache_lifespan
  )
}

# now get the sys deps for git2r for the first time (populating the cache)
pkg_sys_deps("git2r")
#> $libgit2
#> $libgit2$sysreqs
#> [1] "/libgit2/i"
#> 
#> $libgit2$platforms
#> $libgit2$platforms$DEB
#> [1] "libgit2-dev"
#> 
#> $libgit2$platforms$RPM
#> [1] "libgit2-devel"
#> 
#> $libgit2$platforms$PKGBUILD
#> [1] "libgit2"
#> 
#> $libgit2$platforms$`OSX/brew`
#> [1] "libgit2"
#> 
#> 
#> 
#> $openssl
#> $openssl$sysreqs
#> [1] "OpenSSL"
#> 
#> $openssl$platforms
#> $openssl$platforms$DEB
#> [1] "libssl-dev"
#> 
#> $openssl$platforms$PKGBUILD
#> [1] "openssl"
#> 
#> $openssl$platforms$`OSX/brew`
#> [1] "openssl@1.1"
#> 
#> $openssl$platforms$RPM
#> [1] "openssl-devel"
#> 
#> 
#> 
#> $zlib
#> $zlib$sysreqs
#> [1] "/\\bzlib\\b/i"
#> 
#> $zlib$platforms
#> $zlib$platforms$DEB
#> [1] "zlib1g-dev"
#> 
#> $zlib$platforms$`OSX/brew`
#> [1] "zlib"
#> 
#> $zlib$platforms$PKGBUILD
#> [1] "zlib"
#> 
#> $zlib$platforms$RPM
#> [1] "zlib-devel"
#> 
#> 
#> 
#> $libssh2
#> $libssh2$sysreqs
#> [1] "/libssh2/i"
#> 
#> $libssh2$platforms
#> $libssh2$platforms$DEB
#> [1] "libssh2-1-dev"
#> 
#> $libssh2$platforms$RPM
#> [1] "libssh2-devel"
#> 
#> $libssh2$platforms$PKGBUILD
#> [1] "libssh2"
#> 
#> $libssh2$platforms$`OSX/brew`
#> [1] "libssh2"
#> 
#> 
#> 

if (FALSE) {
# for the `cache_lifespan` (we've set a default of 6h), the cached result will be returned
# (as long as `use_cache = TRUE`):
bench::mark("with cache" = pkg_sys_deps("git2r"),
            "without cache" = pkg_sys_deps("git2r", use_cache = FALSE),
            iterations = 10,
            relative = TRUE)}

# purge cache from the above example
pkgpins::board(pkg = "pkgpins") |> pkgpins::purge_cache()